api/requirements_simple.txt
# Минимальные зависимости для Female Task Planner
fastapi==0.115.0
uvicorn==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1

# Google Gemini (совместимые версии)
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.18

# LangChain (упрощенная версия)
langchain-core==0.3.76
langchain-community==0.3.7

# Векторное хранилище (упрощенное)
chromadb==0.5.0
